requirements.txt
Django>=2
djangorestframework>=3.8
celery
django-celery-beat
redis